+++ src/main/java/io/github/adampyramide/BlogAPI/user/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findById(int id);
+    Optional<User> findById(long id);
 
     boolean existsByUsername(String username);
